modified workflow.authorId to type long to support findByAuthorId in workflowRepository fixed Backend getUserWorkflows to return  a list of workflows display list of workflows on profile page

+++ ApacheCMDA-Backend/app/models/WorkflowRepository.java
@@ -29,5 +29,5 @@ public interface WorkflowRepository extends PagingAndSortingRepository<Workflow,
 
 	List<Workflow> findByName(String name);
 	List<Workflow> findById(long id);
-	List<Workflow> findByAuthorId(int id);
+	List<Workflow> findByAuthorId(long id);
 }
